Added radicale onboarding step
