+++ apps/suite-manager/package.json
@@ -15,6 +15,7 @@
     "hono": "^4.12.5",
     "lucide-react": "^0.542.0",
     "pg": "^8.16.3",
+    "qrcode": "^1.5.4",
     "react": "^19.2.4",
     "react-dom": "^19.2.4"
   },
