+++ apps/suite-manager/src/features/onboarding/shared/auth.ts
