+++ apps/suite-manager/src/features/onboarding/shared/state-store.ts
